feat: implement Modules(src/modules) task
